feat: add alwaysRebuild-setting for specific services
